+++ src/components/atoms/Dropdown/index.module.scss
@@ -1,13 +1,13 @@
 @use "src/styles/libs" as *;
+.dropdown_wrapper {
+  min-height: 10px;
+}
 .dropdown {
-  margin-top: 10px;
+  margin: 10px 0;
   width: fit-content;
   background-color: white;
-
   list-style: none;
   box-shadow: 0 4px 10px $BLUISH-GRAY-300, 0 4px 16px $BLUISH-GRAY-100;
-
-  padding: 5px;
   border-radius: 5px;
 }
