+++ src/components/molecules/index.tsx
@@ -5,8 +5,7 @@ export * from "./DragSlider";
 export * from "./ToastContainer";
 export * from "./Footer";
 export * from "./Category";
-export * from "./Selectbox";
 export * from "./Modal";
 export * from "./Table";
-export * from "./DropdownTag";
 export * from "./Input";
+export * from "./Options";
